// src/domains/models/Rider.ts
export class RiderM {
  id: string;
  name: string;
  lastname: string;
  level: number;
  birthdate: Date;
  address: string;
  zipcode: string;
  city: string;
  country: string;
  phone: string;
  createdAt: Date;
  updatedAt: Date;
}

export class RiderForganizationM {
  id: string;
  name: string;
  lastname: string;
  level: number;
  birthdate: Date;
  address: string;
  zipcode: string;
  city: string;
  country: string;
  phone: string;
  createdAt: Date;
  updatedAt: Date;
  status: number;
}